requirements.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg[binary]==3.2.9
dj-database-url==2.1.0

# Caching
redis==5.0.1
django-redis==5.4.0

# API Documentation
drf-spectacular==0.26.5

# API Authentication
djangorestframework-simplejwt==5.2.2
celery==5.3.4

# HTTP Requests
requests==2.31.0

# Environment and Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring and Logging
sentry-sdk==1.38.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0 